% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{Advertising Data (Internal Use)}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{Sales}{Numeric vector representing simulated sales data.}
\item{TV}{Numeric vector representing simulated TV advertising expenses.}
\item{Radio}{Numeric vector representing simulated radio advertising expenses.}
\item{Newspaper}{Numeric vector representing simulated newspaper advertising expenses.}
}
}
\usage{
data(advertising)
}
\description{
A dataset containing simulated advertising expenses and sales figures
over a period of 12 weeks starting from January 1, 2021. This data is
purely fictional and is used for illustrative purposes and development
within the package.
}
\examples{
# This data is for internal development use only and not exported.
data(advertising)
}
\keyword{datasets}
