% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_transformation.R
\name{aggregate_columns}
\alias{aggregate_columns}
\title{Aggregate Columns in a Data Frame}
\usage{
aggregate_columns(modeling_df, aggregated_variables, delimeter = "\\\\|")
}
\arguments{
\item{modeling_df}{A data frame containing the input variables.}

\item{aggregated_variables}{A character vector specifying the columns to aggregate.}

\item{delimeter}{Delimiter used to find individual variables for aggregated_variables.}
}
\value{
A data frame with aggregated values.
}
\description{
This function aggregates specified columns in a data frame and returns a new data frame.
}
\examples{
\dontrun{
  aggregate_columns(advertising, c("Sales", "TV|Radio|Newspaper"), delimeter = "\\\\|")
}

}
