% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{cleanse_model_perfect_fit}
\alias{cleanse_model_perfect_fit}
\title{Cleanse Model Perfect Fit}
\usage{
cleanse_model_perfect_fit(
  lm_model,
  model_data,
  flexible_variables,
  drop_highest_estimate = TRUE,
  ignore_estimate_sign = TRUE
)
}
\arguments{
\item{lm_model}{The linear regression model (of class \code{\link[stats]{lm}})
to be cleansed from variables causing a perfect fit.}

\item{model_data}{The data frame used to fit \code{lm_model}. It must
contain all the variables used in the model.}

\item{flexible_variables}{A character vector specifying the names of
flexible variables in the model. Only these
variables are considered for removal.}

\item{drop_highest_estimate}{Logical; if TRUE, the variable with the highest
(or lowest, if \code{ignore_estimate_sign} is TRUE)
estimate is removed from the model.}

\item{ignore_estimate_sign}{Logical; if TRUE, the absolute value of the
estimates is considered in determining which
variable to drop.}
}
\value{
An updated linear regression model object of the same class as
\code{lm_model}, with variables causing a perfect fit removed.
}
\description{
Removes variables causing a perfect fit in a linear regression model to prevent
issues like overfitting. This is particularly useful in scenarios where a model
might be overly complex due to too many variables or multicollinearity.
}
\examples{
\dontrun{
  advertising <- data.frame(tv = runif(10), radio = runif(10),
   newspaper = runif(10))
  event <- data.frame(event1 = runif(10), event2 = runif(10))
  Sales <- runif(10)

  lm_model <- lm(Sales ~ ., data = cbind(advertising, event))
  cleansed_model <- cleanse_model_perfect_fit(lm_model, model_data = cbind(advertising, event),
    flexible_variables = names(event), drop_highest_estimate = TRUE,
    ignore_estimate_sign = TRUE)
}

}
