% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_iteration.R
\name{collate_models}
\alias{collate_models}
\title{Collate Models}
\usage{
collate_models(
  candidate_variables_list,
  model_df,
  dep_var_info,
  with_intercept = TRUE,
  mdl_start_date = NA,
  mdl_end_date = NA,
  pos_sign_variables,
  neg_sign_variables,
  var_agg_delimiter = "\\\\|",
  apl_delimiter = "_",
  var_apl_delimiter = "\\\\|",
  run_up_to_flexi_vars = 100,
  vif_threshold = 10,
  pvalue_thresholds = c(intercept = 0.15, fixed = 0.15, flexible = 0.15),
  drop_pvalue_precision = 2,
  discard_estimate_sign = TRUE,
  drop_highest_estimate = FALSE,
  get_model_object = FALSE
)
}
\arguments{
\item{candidate_variables_list}{A list of lists, where each sublist contains
candidate predictors for a model.}

\item{model_df}{A dataframe of independent variables for all models.}

\item{dep_var_info}{Information about the dependent variable, such as its
transformation and lag.}

\item{with_intercept}{Logical; if TRUE, an intercept is included in the model.}

\item{mdl_start_date}{The start date to filter the data for model assembly.}

\item{mdl_end_date}{The end date to filter the data for model assembly.}

\item{pos_sign_variables}{A vector of variable names expected to have a positive
sign in the model.}

\item{neg_sign_variables}{A vector of variable names expected to have a negative
sign in the model.}

\item{var_agg_delimiter}{A character string used as the delimiter to aggregate
variables.}

\item{apl_delimiter}{The delimiter used in variable names to separate APL components.}

\item{var_apl_delimiter}{The delimiter used in-between variable names and apl.}

\item{run_up_to_flexi_vars}{An integer representing the number of flexible
variables to consider up to.}

\item{vif_threshold}{A numeric threshold for the Variance Inflation Factor (VIF)
above which variables are dropped.}

\item{pvalue_thresholds}{A named list containing p-value thresholds for
intercept, fixed, and flexible variables.}

\item{drop_pvalue_precision}{The precision for the p-value to drop variables.}

\item{discard_estimate_sign}{Logical; if TRUE, the sign of the estimate is
ignored when dropping variables.}

\item{drop_highest_estimate}{Logical; if TRUE, the variable with the highest
estimate is dropped.}

\item{get_model_object}{Logical; if TRUE, the function returns the linear model
object instead of the default list.}
}
\value{
A list containing the collated models after filtering and preprocessing,
along with associated statistics.
}
\description{
This function extends the \code{collate_base_models} function to accommodate
additional preprocessing of the dependent variable.
}
\examples{
\dontrun{
  # Assuming 'date' is a column in 'independent_df' and 'dependent_df' containing
  # Date class objects.
  filter_start <- as.Date("2000-01-01")
  filter_end <- as.Date("2005-12-31")

  dep_info <- list(transformation = "log", lag = 1)

  results <- collate_models(
    candidate_variables_list,
    model_df = independent_df,
    model_dep_df = dependent_df,
    dep_var_info = dep_info,
    filter_start_date = filter_start,
    filter_end_date = filter_end
    # Additional parameters can be set as needed
  )
}
}
