% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_transformation.R
\name{compose_variable_apl}
\alias{compose_variable_apl}
\title{Compose Variable Names with Adstock, Power, and Lag (APL) Attributes}
\usage{
compose_variable_apl(
  variables_wt_named_apl,
  apl_delimiter = "_",
  delimiter = "|"
)
}
\arguments{
\item{variables_wt_named_apl}{A list where each element is a named vector representing
a variable. The names of the elements are the variable names, and each named
vector contains the APL attributes: adstock, power, and lag.}

\item{apl_delimiter}{The delimiter to use between the variable name and its APL values
(default is "_").}

\item{delimiter}{The delimiter to use between different APL attributes
(default is "|").}
}
\value{
A character vector where each element is a variable name composed with
its APL attributes. These names follow the format:
\verb{variable_name[delimiter]adstock_value[apl_delimiter]power_value[apl_delimiter]lag_value}.
}
\description{
Constructs standardized variable names by incorporating Adstock, Power, and Lag
(APL) attributes into each variable's name. This function is useful for creating
clear and descriptive variable names in models where APL transformations are applied.
}
\examples{
\dontrun{
  variables_wt_named_apl <- list(
    "variable1" = c("adstock" = 0.5, "power" = 2, "lag" = 1),
    "variable2" = c("adstock" = 0.3, "power" = 1, "lag" = 0)
  )
  composed_names <- compose_variable_apl(variables_wt_named_apl,
                                         apl_delimiter = "_",
                                         delimiter = "|")
  print(composed_names)
}

}
