% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{determine_pvalue_flag}
\alias{determine_pvalue_flag}
\title{Determine P-Value Flag}
\usage{
determine_pvalue_flag(type, pvalue, pvalue_thresholds)
}
\arguments{
\item{type}{Character string indicating the type of the predictor (can be 'intercept', 'fixed', or 'flexible').}

\item{pvalue}{Numeric p-value of the predictor.}

\item{pvalue_thresholds}{Named numeric vector of thresholds for each predictor type. Names can be 'intercept', 'fixed', and 'flexible'.}
}
\value{
Logical; TRUE if the p-value is above the threshold for its type.
}
\description{
Flags p-values based on predefined thresholds for different types of predictors.
}
