% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{determine_vif_flag}
\alias{determine_vif_flag}
\title{Determine VIF Flag}
\usage{
determine_vif_flag(type, vif, vif_threshold)
}
\arguments{
\item{type}{Character string indicating the type of the variable (either 'fixed' or 'flexible').}

\item{vif}{Numeric VIF value for the predictor.}

\item{vif_threshold}{Threshold above which the VIF flag should be triggered.}
}
\value{
Logical; TRUE if the VIF is greater than the threshold. If there is only one 'fixed' variable and the VIF is infinite, then it will return FALSE.
}
\description{
This function determines the VIF flag for predictors based on their type and given VIF value.
}
