% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_transformation.R
\name{get_dep_indep_vars}
\alias{get_dep_indep_vars}
\title{Get Dependent and Independent Variables}
\usage{
get_dep_indep_vars(
  model_variable,
  var_agg_delimiter = "|",
  trim = TRUE,
  print_model_type = TRUE
)
}
\arguments{
\item{model_variable}{A character string representing the model variable,
typically a formula or expression where variables may be aggregated or
segregated using '+' and '-'.}

\item{var_agg_delimiter}{A character string representing the delimiter for
aggregated variables. Default is "|".}

\item{trim}{A logical indicating whether to trim whitespace from the output.
Default is TRUE.}

\item{print_model_type}{A logical indicating whether to print the determined
model type to the console. Default is TRUE.}
}
\value{
A list with two elements: 'dependent_var' containing the processed
dependent variable(s) and 'independent_var' containing the processed
independent variable(s).
}
\description{
Determines the type of a model based on the presence of characters '+' and '-',
and processes the model variable to extract dependent and independent variables.
It classifies models into types like "Remodel", "Aggregate", "Segregate",
or "Aggregate & Segregate" based on the syntax of the model variable.
}
\examples{
\dontrun{
  # Aggregation
  get_dep_indep_vars("TV_0_1_0+Sales_0_1_0")
  # Remodel
  get_dep_indep_vars("TV_0_1_0")
  # Segregation
  get_dep_indep_vars("TV-Sales_0_1_0")
  # Aggregation & Segregation
  get_dep_indep_vars("TV_0_1_0+-Sales_0_1_0")
  # Aggregation & Segregation, segregation is evaluated first and then aggregation
  get_dep_indep_vars("TV_0_1_0+Sales-Radio_0_1_0")
  # Aggregation with alternate delimiter
  get_dep_indep_vars("TV_0_1_0+Sales|Radio_0_1_0", var_agg_delimiter = "|")
}

}
