% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_transformation.R
\name{parse_variable_wt_apl}
\alias{parse_variable_wt_apl}
\title{Parse Variable with Adstock, Power, and Lag}
\usage{
parse_variable_wt_apl(
  variables_wt_apl,
  apl_delimeter = "_",
  delimeter = "\\\\|"
)
}
\arguments{
\item{variables_wt_apl}{A character vector with each element containing variable names and up to three numeric values separated by \code{delimeter} and \code{apl_delimeter}.}

\item{apl_delimeter}{A string representing the delimiter between adstock, power, and lag values.}

\item{delimeter}{A string representing the main delimiter between variable and apl in the variables_wt_apl strings.}
}
\value{
A data frame containing variables and their corresponding adstock, power, and lag values.
}
\description{
This function parses a vector of strings containing variable information along with adstock, power, and lag, separated by specified delimiters.
Each string should contain variable names and up to three numeric values (adstock, power, and lag) in a consistent order, separated by delimiters.
}
\examples{
\dontrun{
parse_variable_wt_apl(c("TV_Smart|0.8_0.22_0.11", "Radio|0.5_0.15"), "_", "\\\\|")
}
}
