% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{calculate_vif}
\alias{calculate_vif}
\title{Calculate Variance Inflation Factor (VIF)}
\usage{
calculate_vif(model)
}
\arguments{
\item{model}{A \code{lm} or \code{glm} object representing the fitted linear regression model.}
}
\value{
A numeric vector containing the VIF values for each predictor variable in the model.
}
\description{
This function calculates the VIF for a given linear regression model. It returns a numeric vector containing the VIF values for each predictor variable. If the calculation fails (e.g., due to multicollinearity issues), it returns \code{Inf} for all predictor variables.
}
\examples{
\dontrun{
  lm_model <- lm(Sales ~ ., data = advertising)
  calculate_vif(lm_model)
}

}
