% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{cleanse_model_perfect_fit}
\alias{cleanse_model_perfect_fit}
\title{Cleanse Model Perfect Fit}
\usage{
cleanse_model_perfect_fit(
  lm_model,
  model_data,
  flexible_variables,
  drop_highest_estimate = TRUE,
  ignore_estimate_sign = TRUE
)
}
\arguments{
\item{lm_model}{The linear regression model to be cleansed.}

\item{model_data}{Data used for modeling i.e. lm_model.}

\item{flexible_variables}{A character vector specifying the flexible variables
in the model.}

\item{drop_highest_estimate}{Logical indicating whether to drop the variable
with the highest estimate.}

\item{ignore_estimate_sign}{Logical indicating whether to ignore the sign of
the estimate while dropping variables.}
}
\value{
A linear regression model with variables causing a perfect fit
removed.
}
\description{
This function removes variables causing a perfect fit in a linear regression
model, which can lead to issues like over-fitting.
}
\examples{
\dontrun{
  lm_model <- lm(Sales ~ ., data = cbind(advertising, event))
  cleanse_model_perfect_fit(lm_model, model_data = cbind(advertising, event),
    flexible_variables = names(event), drop_highest_estimate = TRUE,
    ignore_estimate_sign = TRUE)
}

}
