% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{cleanse_model_singularity}
\alias{cleanse_model_singularity}
\title{Cleanse Model Singularity}
\usage{
cleanse_model_singularity(
  lm_model,
  model_data,
  flexible_variables,
  round_digits = 2,
  verbose = FALSE
)
}
\arguments{
\item{lm_model}{The linear regression model to be cleansed.}

\item{model_data}{Data used for modeling i.e. lm_model}

\item{flexible_variables}{A character vector specifying the flexible variables in the model.}

\item{round_digits}{Number of digits to round to identify order of variables dropped with singular dependencies.}

\item{verbose}{Logical; if TRUE, prints the variable being dropped in each iteration.}
}
\value{
A linear regression model with singular variables removed.
}
\description{
This function removes singular variables from a linear regression model to improve model stability.
}
\examples{
\dontrun{
  lm_model <- lm(Sales ~ ., data = cbind(advertising, event))
  cleanse_model_singularity(lm_model, names(event), round_digits = 2, verbose = TRUE)
}

}
