% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_iteration.R
\name{collate_base_models}
\alias{collate_base_models}
\title{Collate Base Models}
\usage{
collate_base_models(
  candidate_variables_list,
  model_df,
  model_dep_df,
  with_intercept = TRUE,
  pos_sign_variables,
  neg_sign_variables,
  var_agg_delimiter = "\\\\|",
  run_up_to_flexi_vars = 0,
  vif_threshold = 10,
  pvalue_thresholds = c(intercept = 0.15, fixed = 0.15, flexible = 0.15),
  drop_pvalue_precision = 2,
  discard_estimate_sign = TRUE,
  drop_highest_estimate = FALSE,
  get_model_object = FALSE
)
}
\arguments{
\item{candidate_variables_list}{A list of lists, where each sublist contains
candidate predictors for a model.}

\item{model_df}{A dataframe of independent variables for all models.}

\item{model_dep_df}{A dataframe of the dependent variable 'Y' for all models.}

\item{with_intercept}{Logical; if TRUE, an intercept is included in the model.}

\item{pos_sign_variables}{A vector of variable names expected to have a positive
sign in the model.}

\item{neg_sign_variables}{A vector of variable names expected to have a negative
sign in the model.}

\item{var_agg_delimiter}{A character string used as the delimiter to aggregate
variables.}

\item{run_up_to_flexi_vars}{An integer representing the number of flexible
variables to consider up to.}

\item{vif_threshold}{A numeric threshold for the Variance Inflation Factor (VIF)
above which variables are dropped.}

\item{pvalue_thresholds}{A named list containing p-value thresholds for
intercept, fixed, and flexible variables.}

\item{drop_pvalue_precision}{The precision for the p-value to drop variables.}

\item{discard_estimate_sign}{Logical; if TRUE, the sign of the estimate is
ignored when dropping variables.}

\item{drop_highest_estimate}{Logical; if TRUE, the variable with the highest
estimate is dropped.}

\item{get_model_object}{Logical; if TRUE, the function returns the linear model
object instead of the default list.}
}
\value{
A list containing the collated base models and associated statistics.
}
\description{
This function collates base models for linear regression from a list of candidate
variable sets, using the \code{assemble_base_models} function for model assembly.
}
\examples{
\dontrun{
  candidate_variables_list <- list(
    list(fixed = c("var1"), flexible = c("var3")),
    list(fixed = c("var2"), flexible = c("var4"))
  )
  independent_df <- data.frame(var1 = rnorm(100), var2 = rnorm(100))
  dependent_df <- data.frame(Y = rnorm(100))

  results <- collate_base_models(
    candidate_variables_list,
    model_df = independent_df,
    model_dep_df = dependent_df
    # Additional parameters can be set as needed
  )
}
}
