% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_transformation.R
\name{compose_variable_apl}
\alias{compose_variable_apl}
\title{Compose Variable Names with Adstock, Power, and Lag (APL) Attributes}
\usage{
compose_variable_apl(
  variables_wt_named_apl,
  apl_delimiter = "_",
  delimiter = "|"
)
}
\arguments{
\item{variables_wt_named_apl}{A list where each element is named after a variable and contains a
named vector of APL attributes: adstock, power, and lag.}

\item{apl_delimiter}{The delimiter to use between the variable name and APL values.}

\item{delimiter}{The delimiter to use between APL attributes.}
}
\value{
A character vector of variable names composed with their APL attributes.
}
\description{
This function takes a list of variables with their corresponding Adstock, Power, and Lag (APL)
attributes and composes a standardized naming convention for each variable incorporating
its APL attributes.
}
\examples{
\dontrun{
  variables_wt_named_apl <- list(
    "variable1" = c("adstock" = 0.5, "power" = 2, "lag" = 1),
    "variable2" = c("adstock" = 0.3, "power" = 1, "lag" = 0)
  )
  composed_names <- compose_variable_apl(variables_wt_named_apl)
}

}
