% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_transformation.R
\name{compute_apl_values}
\alias{compute_apl_values}
\title{Apply Adstock, Power, and Lag Transformation to a Data Frame}
\usage{
compute_apl_values(modeling_df, adstock, power, lag)
}
\arguments{
\item{modeling_df}{A data frame containing the input variables.}

\item{adstock}{The adstock rate for the transformation.}

\item{power}{The power transformation to capture the non-linear effectiveness of marketing activity.}

\item{lag}{The lag parameter for the time lag in the transformation.}
}
\value{
A data frame with Adstock, Power and Lag transformed values.
}
\description{
This function applies the Adstock, Power, and Lag transformation to a given data frame.
}
\details{
The function applies the Adstock transformation using a recursive filtering
method and raises the filtered values to the power specified. It then lags
the result by the specified time lag.
}
\examples{
\dontrun{
  compute_apl_values(advertising, adstock = 0.8, power = 2, lag = 1)
  compute_apl_values(advertising[,"Sales",drop=F], adstock = 0.8, power = 2, lag = 1)
}

}
