% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_transformation.R
\name{decompose_model_component}
\alias{decompose_model_component}
\title{Decompose Model Component with Adstock, Power, and Lag (APL) Transformations}
\usage{
decompose_model_component(
  variables_wt_weights,
  model_df,
  is_weight_coefficient = TRUE,
  apl_delimiter = "_",
  delimiter = "\\\\|"
)
}
\arguments{
\item{variables_wt_weights}{A named vector of weights for the variables, where names include APL info.}

\item{model_df}{Data frame containing the variables to be transformed.}

\item{is_weight_coefficient}{Logical indicating if the weights should be treated as coefficients.}

\item{apl_delimiter}{The delimiter used in variable names to separate APL components.}

\item{delimiter}{The delimiter used in variable names to denote different variables.}
}
\value{
A transformed data frame with variables weighted and APL transformations applied.
}
\description{
This function takes a set of variables with weights and applies Adstock, Power,
and Lag (APL) transformations to the variables within a model dataframe.
It handles the decomposition of model components based on provided coefficients
or contributions.
}
\examples{
\dontrun{
  advertising <- data.frame(TV = rnorm(100), Radio = rnorm(100))
  event <- data.frame(Date2021_01_08 = sample(0:1, 100, replace = TRUE),
                      Date2021_01_01 = sample(0:1, 100, replace = TRUE))
  model_df <- cbind(advertising, event)

  variables_wt_weights <- setNames(1:5, c("TV|0.8_0.22_0", "Radio|0.5_0.15_1",
                                          "Intercept", "Date2021_01_08", "Date2021_01_01"))

  transformed_df <- decompose_model_component(variables_wt_weights, model_df,
                                              is_weight_coefficient = FALSE,
                                              apl_delimiter = "_",
                                              delimiter = "\\\\|")
}
}
