% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{determine_expected_sign}
\alias{determine_expected_sign}
\title{Determine Expected Sign of Coefficients}
\usage{
determine_expected_sign(variable, pos_vars, neg_vars, var_agg_delimiter)
}
\arguments{
\item{variable}{Character string of variable names.}

\item{pos_vars}{Character vector of variables expected to have a positive sign.}

\item{neg_vars}{Character vector of variables expected to have a negative sign.}

\item{var_agg_delimiter}{Character string delimiter used in variable names for aggregation.}
}
\value{
Logical; TRUE if the variable is expected to have a positive sign, FALSE if expected to have a negative sign, NA otherwise.
}
\description{
Determines the expected sign (positive or negative) of model coefficients.
}
