% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_transformation.R
\name{generate_apl_combinations}
\alias{generate_apl_combinations}
\title{Generate Adstock, Power, and Lag Combinations}
\usage{
generate_apl_combinations(
  adstock_start_end_step,
  power_start_end_step,
  lag_start_end_step,
  apl_constraints = NA
)
}
\arguments{
\item{adstock_start_end_step}{A named numeric vector specifying the start, end, and step for adstock.}

\item{power_start_end_step}{A named numeric vector specifying the start, end, and step for power.}

\item{lag_start_end_step}{A named numeric vector specifying the start, end, and step for lag.}

\item{apl_constraints}{A list containing additional constraints for adstock, power, and lag.}
}
\value{
A list of combinations for adstock, power, and lag.
}
\description{
This function generates combinations of adstock, power, and lag based on specified constraints.
}
\examples{
\dontrun{
  variable_apl_constraints <- list(
    TV = list(
      adstock = setNames(c(.1, .3, .1), c("start", "end", "step")),
      power = setNames(c(.2, .5, .1), c("start", "end", "step")),
      lag = setNames(c(0, 2, 1), c("start", "end", "step")),
      constraints = "adstock <= power"
    )
  )
  apl_combinations <- generate_apl_combinations(
    variable_apl_constraints$TV["adstock"],
    variable_apl_constraints$TV["power"],
    variable_apl_constraints$TV["lag"],
    variable_apl_constraints$TV["constraints"]
  )
  print(apl_combinations)
}

}
