% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_transformation.R
\name{generate_apl_dataframe}
\alias{generate_apl_dataframe}
\title{Apply Adstock, Power, and Lag Vectorized Transformation to a Data Frame}
\usage{
generate_apl_dataframe(
  modeling_df,
  adstock,
  power,
  lag,
  apl_delimeter = "_",
  delimeter = "|"
)
}
\arguments{
\item{modeling_df}{A data frame containing the input variables.}

\item{adstock}{A vector of adstock rates for the transformations.}

\item{power}{A vector of power transformations to capture the non-linear effectiveness of marketing activity.}

\item{lag}{A vector of lag parameters for the time lag in the transformations.}

\item{apl_delimeter}{Delimiter used to separate Adstock, Power, and Lag values in resulting column names.}

\item{delimeter}{Delimiter used to separate original column name and apl in output data frame.}
}
\value{
A data frame with vectorized Adstock, Power, and Lag-transformed values.
}
\description{
This function applies vectorized Adstock, Power, and Lag transformations to a given data frame.
}
\examples{
\dontrun{
  generate_apl_dataframe(advertising, adstock = c(0.8, 0.9), power = c(2, 3), lag = c(1, 2))
}

}
