% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_transformation.R
\name{generate_variable_combination}
\alias{generate_variable_combination}
\title{Generate Variable Combinations with Adstock, Power, and Lag}
\usage{
generate_variable_combination(variables_wt_apl_constraints)
}
\arguments{
\item{variables_wt_apl_constraints}{A list containing named vectors for each variable with adstock, power, lag, and constraints.

Each named vector should include the following components:
\describe{
\item{adstock}{A named numeric vector specifying the start, end, and step for adstock.}
\item{power}{A named numeric vector specifying the start, end, and step for power.}
\item{lag}{A named numeric vector specifying the start, end, and step for lag.}
\item{constraints}{An optional character string specifying additional constraints for adstock and power.}
}}
}
\value{
A list of combinations for adstock, power, and lag for each variable.
}
\description{
This function generates combinations of adstock, power, and lag for multiple variables based on specified constraints.
}
\examples{
\dontrun{
  variables_wt_apl_constraints <- list(
    TV = list(
      adstock = setNames(c(.1, .3, .1), c("start", "end", "step")),
      power = setNames(c(.2, .5, .1), c("start", "end", "step")),
      lag = setNames(c(0, 2, 1), c("start", "end", "step")),
      constraints = "adstock <= power"
    ),
    Radio = list(
      adstock = setNames(c(.2, .5, .1), c("start", "end", "step")),
      power = setNames(c(.4, .7, .1), c("start", "end", "step")),
      lag = setNames(c(0, 0, 1), c("start", "end", "step")),
      constraints = "adstock <= power"
    )
  )
  generate_variable_combination(variables_wt_apl_constraints)
}

}
