% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{get_base_model}
\alias{get_base_model}
\title{Get Base Model After Iteratively Dropping Variables}
\usage{
get_base_model(
  lm_model,
  model_data,
  independent_var_info,
  pos_vars,
  neg_vars,
  var_agg_delimiter = "\\\\|",
  run_up_to_flexi_vars = 10,
  vif_threshold = 10,
  pvalue_thresholds = c(intercept = 0.15, fixed = 0.15, flexible = 0.15),
  drop_pvalue_precision = 2,
  discard_estimate_sign = TRUE,
  drop_highest_estimate = FALSE,
  get_model_object = FALSE
)
}
\arguments{
\item{lm_model}{A linear model object (of class \code{\link[stats]{lm}}) to be processed.}

\item{model_data}{Data used for fitting \code{lm_model}.}

\item{independent_var_info}{A data frame with information about independent
variables in the model. It should contain columns for variable names, adstock,
power, lag, and the type of variable (e.g., "flexible", "fixed").}

\item{pos_vars}{Character vector of variables expected to have a positive
relationship with the response.}

\item{neg_vars}{Character vector of variables expected to have a negative
relationship with the response.}

\item{var_agg_delimiter}{Delimiter used in variable names for aggregation
(default is "\\|").}

\item{run_up_to_flexi_vars}{Integer indicating the number of "flexible"
variables to consider for retention (default is 10).}

\item{vif_threshold}{Threshold for VIF above which variables are considered
for removal (default is 10).}

\item{pvalue_thresholds}{Named numeric vector of p-value thresholds for
different types of variables (default is c(intercept = 0.15, fixed = 0.15, flexible = 0.15)).}

\item{drop_pvalue_precision}{Integer for rounding p-values in the decision
process (default is 2).}

\item{discard_estimate_sign}{Logical; if TRUE, the sign of estimates is
disregarded in the decision process (default is TRUE).}

\item{drop_highest_estimate}{Logical; if TRUE, always drop the variable
with the highest estimate (default is FALSE).}

\item{get_model_object}{Logical; if TRUE, returns the final model object
(default is FALSE).}
}
\value{
A list containing the data frame of model coefficients, the data frame
of model summary statistics, and optionally the final model object if
\code{get_model_object} is TRUE.
}
\description{
Iteratively updates a linear model by dropping variables based on criteria such as
high p-values, high VIF, and expected sign of coefficients. The process continues
until no more variables meet the criteria for exclusion, refining the model.
}
\examples{
\dontrun{
  advertising <- data.frame(tv = runif(10), radio = runif(10), newspaper = runif(10),
   sales = runif(10))
  event <- data.frame(event1 = runif(10), event2 = runif(10))
  lm_model <- lm(sales ~ tv + radio + newspaper + event1 + event2, data = cbind(advertising, event))
  independent_var_info <- data.frame(variable = c("tv", "radio", "newspaper", "event1", "event2"),
                                     type = c("fixed", "fixed", "fixed", "flexible", "flexible"),
                                     adstock = rep(NA, 5), power = rep(NA, 5), lag = rep(NA, 5))
  base_model <- get_base_model(lm_model, cbind(advertising, event), independent_var_info,
                               pos_vars = c("tv", "radio"), neg_vars = c("newspaper"),
                               get_model_object = TRUE)
}

}
