% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{identify_drop_variable}
\alias{identify_drop_variable}
\title{Identify Variable to Drop from the Model}
\usage{
identify_drop_variable(
  coef_df,
  pvalue_precision,
  discard_sign,
  highest_estimate,
  run_up_to_flexi_vars
)
}
\arguments{
\item{coef_df}{Data frame of model coefficients.}

\item{pvalue_precision}{Integer; number of decimal places to round p-values to.}

\item{discard_sign}{Logical; whether to discard the sign of estimates in the decision process.}

\item{highest_estimate}{Logical; whether to consider the highest estimate for dropping.}

\item{run_up_to_flexi_vars}{Integer; up to how many flexible variables to consider for keeping in the final model.}
}
\value{
Character; the name of the variable suggested to be dropped.
}
\description{
Identifies which variable should be dropped from the model based on p-value and estimation criteria.
}
