% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_transformation.R
\name{parse_variable_wt_apl}
\alias{parse_variable_wt_apl}
\title{Parse Variable with Adstock, Power, and Lag}
\usage{
parse_variable_wt_apl(variables_wt_apl, apl_delimiter = "_", delimiter = "|")
}
\arguments{
\item{variables_wt_apl}{A character vector where each element contains a
variable name and up to three numeric values for adstock, power, and lag,
separated by \code{delimiter} and \code{apl_delimiter}.}

\item{apl_delimiter}{A string representing the delimiter used between the
adstock, power, and lag values in each string. It cann't be period (.)}

\item{delimiter}{A string representing the main delimiter used between the
variable name and the APL attributes in each string. It cann't be period (.)}
}
\value{
A data frame where each row corresponds to an element in
\code{variables_wt_apl}, with columns for the variable name, adstock value,
power value, and lag value.
}
\description{
Parses a vector of strings containing variable information along with Adstock,
Power, and Lag (APL) attributes, separated by specified delimiters. Each string
in the vector should contain a variable name followed by up to three numeric
values representing the APL attributes, in a consistent order and separated
by the delimiters.
}
\examples{
\dontrun{
  parsed_data <- parse_variable_wt_apl(c("TV_Smart|0.8_0.22_0.11", "Radio|0.5_0.15"),
                                       apl_delimiter = "_",
                                       delimiter = "|")
  print(parsed_data)
}

}
