% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{update_model}
\alias{update_model}
\title{Update Linear Model by Dropping a Variable}
\usage{
update_model(model, model_data, variable_to_drop)
}
\arguments{
\item{model}{An object of class \code{\link[stats]{lm}} or similar, representing
the fitted model to be updated.}

\item{model_data}{A dataframe containing the data that the model was originally
fitted with. It must contain all the necessary columns used by the model.}

\item{variable_to_drop}{A character string indicating the name of the variable
to be dropped from the model. If "(Intercept)" is provided, the model will be
updated without an intercept.}
}
\value{
An object of the same class as \code{model}, representing the updated
model without the specified variable or intercept.
}
\description{
This function updates a given linear model by removing a specified variable.
The model is re-fitted without the specified variable or the intercept if
"(Intercept)" is passed as the variable to drop.
}
\examples{
\dontrun{
  data(mtcars)
  initial_model <- lm(mpg ~ wt + cyl + gear, data = mtcars)
  updated_model <- update_model(initial_model, mtcars, "gear")

  # To remove the intercept:
  updated_model_no_intercept <- update_model(initial_model, mtcars, "(Intercept)")
}
}
